+++ src/lib/app-provider/hooks/useExampleAddresses.ts
@@ -1,4 +1,4 @@
-import { toBech32 } from "@cosmjs/encoding";
+import { toBech32, toHex } from "@cosmjs/encoding";
 import { useMemo } from "react";
 
 import type { BechAddr20, BechAddr32 } from "lib/types";
@@ -10,23 +10,23 @@ export const useExampleAddresses = () => {
   const { bech32Prefix } = useCurrentChain();
 
   const generateExampleAddresses = () => {
-    const bytes = Array.from(Array(32).keys());
-    const user = toBech32(
-      bech32Prefix,
-      Uint8Array.from(bytes.slice(0, 20))
-    ) as BechAddr20;
+    const bytes32 = Array.from(Array(32).keys());
+    const bytes20 = bytes32.slice(0, 20);
+    const user = toBech32(bech32Prefix, Uint8Array.from(bytes20)) as BechAddr20;
 
     // reverse the bytes so the initial characters are different from the user address
-    const contract = toBech32(
+    const evmContract = "0x" + toHex(Uint8Array.from(bytes20).reverse());
+    const wasmContract = toBech32(
       bech32Prefix,
-      Uint8Array.from(bytes.reverse())
+      Uint8Array.from(bytes32.reverse())
     ) as BechAddr32;
 
     const validator = addrToValoper(user);
 
     return {
       user,
-      contract,
+      wasmContract,
+      evmContract,
       validator,
     };
   };