+++ src/lib/pages/evm-contract-verify/components/solidity/EvmContractVerifySolidityUploadFiles.tsx
@@ -1,23 +1,3 @@
-import { EvmContractVerifyForm } from "../../types";
-import { Control } from "react-hook-form";
-import { ConstructorArgs } from "../ConstructorArgs";
-import { OptimizerConfiguration } from "../OptimizerConfiguration";
-
-interface EvmContractVerifySolidityUploadFilesProps {
-  control: Control<EvmContractVerifyForm>;
-}
-
-export const EvmContractVerifySolidityUploadFiles = ({
-  control,
-}: EvmContractVerifySolidityUploadFilesProps) => {
-  return (
-    <div>
-      TODO: EvmContractVerifySolidityUploadFiles
-      <br />
-      <br />
-      <ConstructorArgs control={control} />
-      <br />
-      <OptimizerConfiguration control={control} />
-    </div>
-  );
+export const EvmContractVerifySolidityUploadFiles = () => {
+  return <div>TODO: EvmContractVerifySolidityUploadFiles</div>;
 };